backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
PyJWT>=2.8.0
openai>=1.12.0
langchain>=0.1.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchainhub>=0.1.0
langgraph>=1.0.0
langsmith>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.1




